backend-python/requirements.txt
# Backend Python - RoboCup Middleware

flask>=2.3.0
flask-socketio>=5.3.0
paho-mqtt>=1.6.0
pytest>=7.4.0
python-dotenv>=1.0.0
eventlet>=0.33.0
